Update to Slate 0.34
