+++ package.json
@@ -9,7 +9,7 @@
     "react": "16.0.0",
     "react-dom": "16.0.0",
     "react-scripts": "1.0.14",
-    "slate": "0.33.6",
+    "slate": "0.34.0",
     "slate-react": "0.12.4",
     "slate-edit-list": "0.11.3"
   },
